+++ include/vsmc/rng/threefry_avx2.hpp
@@ -76,7 +76,7 @@ struct ThreefryCtrPackAVX2 {
     static void eval(ctr_type &ctr, state_type &state)
     {
         ctr_block_type ctr_block;
-        Counter<ctr_type>::increment(ctr, ctr_block);
+        internal::increment(ctr, ctr_block);
         pack<0>(ctr_block, state, std::integral_constant<bool, 0 < K>());
     }
 
@@ -136,10 +136,6 @@ class ThreefryEngineAVX2
     typedef std::array<ResultType, K> ctr_type;
     typedef std::array<ResultType, K> key_type;
 
-    private:
-    typedef Counter<ctr_type> counter;
-
-    public:
     explicit ThreefryEngineAVX2(result_type s = 0) : index_(M_)
     {
         VSMC_STATIC_ASSERT_RNG_THREEFRY(AVX2);
@@ -165,7 +161,7 @@ class ThreefryEngineAVX2
 
     void seed(result_type s)
     {
-        counter::reset(ctr_);
+        ctr_.fill(0);
         key_type k;
         k.fill(0);
         k.front() = s;
@@ -179,7 +175,7 @@ class ThreefryEngineAVX2
             key_type, ThreefryEngineAVX2<ResultType, K, Rounds>>::value>::type
             * = nullptr)
     {
-        counter::reset(ctr_);
+        ctr_.fill(0);
         key_type k;
         seq.generate(k.begin(), k.end());
         init_par(k);
@@ -188,7 +184,7 @@ class ThreefryEngineAVX2
 
     void seed(const key_type &k)
     {
-        counter::reset(ctr_);
+        ctr_.fill(0);
         init_par(k);
         index_ = M_;
     }
@@ -206,7 +202,7 @@ class ThreefryEngineAVX2
 
     void ctr(const ctr_type &c)
     {
-        counter::set(ctr_, c);
+        ctr_ = c;
         index_ = M_;
     }
 
@@ -242,7 +238,7 @@ class ThreefryEngineAVX2
             return;
         }
 
-        counter::increment(ctr_, static_cast<result_type>(n / M_));
+        internal::increment(ctr_, static_cast<result_type>(n / M_));
         index_ = M_;
         operator()();
         index_ = n % M_;